+++ tests/at_onboarding_cli_functional_tests_proxy/lib/at_demo_data_wrapper.dart
@@ -1,0 +1,8 @@
+import 'package:at_demo_data/at_demo_data.dart';
+
+const int atSignIndex = 30; // @gary
+
+final Map<String, String> atSignData = {
+  'atSign': allAtsigns[atSignIndex],
+  'cramKey': cramKeyMap[allAtsigns[atSignIndex]] ?? '',
+};