+++ tests/at_onboarding_cli_functional_tests_proxy/restart.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+sudo docker-compose down
+sudo docker-compose up -d